+++ sdk/python/pulumi_digitalocean/floating_ip_assignment.py
@@ -5,7 +5,7 @@
 import warnings
 import pulumi
 import pulumi.runtime
-from typing import Any, Dict, List, Mapping, Optional, Tuple, Union
+from typing import Any, Mapping, Optional, Sequence, Union
 from . import _utilities, _tables
 
 __all__ = ['FloatingIpAssignment']
@@ -15,7 +15,7 @@ class FloatingIpAssignment(pulumi.CustomResource):
     def __init__(__self__,
                  resource_name: str,
                  opts: Optional[pulumi.ResourceOptions] = None,
-                 droplet_id: Optional[pulumi.Input[float]] = None,
+                 droplet_id: Optional[pulumi.Input[int]] = None,
                  ip_address: Optional[pulumi.Input[str]] = None,
                  __props__=None,
                  __name__=None,
@@ -45,7 +45,7 @@ class FloatingIpAssignment(pulumi.CustomResource):
 
         :param str resource_name: The name of the resource.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[float] droplet_id: The ID of Droplet that the Floating IP will be assigned to.
+        :param pulumi.Input[int] droplet_id: The ID of Droplet that the Floating IP will be assigned to.
         :param pulumi.Input[str] ip_address: The Floating IP to assign to the Droplet.
         """
         if __name__ is not None:
@@ -81,7 +81,7 @@ class FloatingIpAssignment(pulumi.CustomResource):
     def get(resource_name: str,
             id: pulumi.Input[str],
             opts: Optional[pulumi.ResourceOptions] = None,
-            droplet_id: Optional[pulumi.Input[float]] = None,
+            droplet_id: Optional[pulumi.Input[int]] = None,
             ip_address: Optional[pulumi.Input[str]] = None) -> 'FloatingIpAssignment':
         """
         Get an existing FloatingIpAssignment resource's state with the given name, id, and optional extra
@@ -90,7 +90,7 @@ class FloatingIpAssignment(pulumi.CustomResource):
         :param str resource_name: The unique name of the resulting resource.
         :param pulumi.Input[str] id: The unique provider ID of the resource to lookup.
         :param pulumi.ResourceOptions opts: Options for the resource.
-        :param pulumi.Input[float] droplet_id: The ID of Droplet that the Floating IP will be assigned to.
+        :param pulumi.Input[int] droplet_id: The ID of Droplet that the Floating IP will be assigned to.
         :param pulumi.Input[str] ip_address: The Floating IP to assign to the Droplet.
         """
         opts = pulumi.ResourceOptions.merge(opts, pulumi.ResourceOptions(id=id))
@@ -103,7 +103,7 @@ class FloatingIpAssignment(pulumi.CustomResource):
 
     @property
     @pulumi.getter(name="dropletId")
-    def droplet_id(self) -> pulumi.Output[float]:
+    def droplet_id(self) -> pulumi.Output[int]:
         """
         The ID of Droplet that the Floating IP will be assigned to.
         """
